+++ src/redux/actions/provider-packages/tests/get-provider-packages.test.js
@@ -1,0 +1,13 @@
+import {
+  getProviderPackages,
+  GET_PROVIDER_PACKAGES,
+} from '../get-provider-packages';
+
+describe('getProviderPackagesAction', () => {
+  it('should return an action object', () => {
+    expect(getProviderPackages('id')).toEqual({
+      type: GET_PROVIDER_PACKAGES,
+      payload: 'id',
+    });
+  });
+});